Moved cluster view to notebook module.
